moving hashstring to common
